Maintain a Swift type name -> imported type mapping in jextract-swift
Switch the representation of the imported types from a flat list over to
a dictionary mapping from the Swift name to the imported type. We need
this so that we can find information about the Swift type when it is
referenced elsewhere in the syntax tree, such as extensions or uses of
that type in other function signatures.

While here, stop recording intermediate results within the visitor
itself. Instead, keep a reference to the full translator and record
results there incrementally. This will become more important when we
start handling extensions.