Introduce the @JavaStaticMethod and use it for calling static Java methods
The use of the `@JavaMethod` macro within an extension of `JavaClass` to
represent static Java methods on that class didn't quite work, because
JavaClass is itself an AnyJavaObject with its own methods. Introduce
a `@JavaStaticMethod` macro that will be used for this purpose, and
have it call through a separate API (`dynamicJavaStaticMethodCall`).

Thank you to Lokesh for reporting this bug!