Temporarily switch JavaKitTests target to XCTest
Command-line testing on macOS is broken for the JavaKitTests due to an
interaction between the Java Virtual Machine, a helper test process
for swift-testing (`swiftpm-testing-helper`), and some macOS security
features. Work around the issue by replacing our use of swift-testing
with XCTest, whose helper process doesn't seem to have this problem.
Only do so for test targets that need to spin up an instance of the
JVM, and keep using swift-testing everywhere we can.

This makes me very sad, but is better than the current solution of
disabling these tests when running from the terminal. Once we've found
a solution for the underlying issue, we'll revert this change as get
back to swift-testing and its much nicer testing output.

Fixes issue #43 without disabling tests.