Lazily attach threads to the JVM when querying the JNI environment
Instead of having the JavaVirtualMachine instance keep track of the JNI
environment for the thread in which it was initialized, provide a JNI
environment query function that attaches the current thread to the JVM
if needed. This lets us query the JNI environment for multiple threads,
should we want to.

We no longer need to force all of the JavaKit runtime tests to the main
actor.