Introduce support for a shared JavaVirtualMachine instance
Java implementations have a longstanding limitation that there may only
be one Java Virtual Machine instance per process. Take advantage of this
to provide a shared JavaVirtualMachine instance, initialized on first
use, that can be accessed from any thread. This eliminates the need
to force JavaVirtualMachine creation to be on the main actor, since we
can do it from any thread and get the same result.