Add a SwiftPM build plugin to generate Swift wrappers for Java classes
Introduce the plugin `Java2SwiftPlugin`, which will invoke Java2Swift to
generate Swift wrappers for the specified Java classes. Each SwiftPM
target that uses this plugin must contain a file named `Java2Swift.config`
that specifies options for the Java2Swift tool. This is a JSON file
with a top-level dictionary that currently can contain keys for the
Java class path (`classPath`) and a dictionary called `classes` whose
keys are Java canonical class names (such as `java.util.Vector`) and
whose values are the corresponding Swift names. For example:

```swift
{
  "classes" : {
    "java.util.Vector" : "JavaVector"
  }
}

```

will translate the class `java.util.Vector` into the Swift type `JavaVector`.
The resulting Swift files will become part of the Swift module that uses
this plugin.

I've updated the JavaKitSampleApp to make use of this, albeit in a very
narrow way that is not at all compelling. The user guide also needs updating.

Fixes issue #16