Java2Swift: Produce a protocol to aid with native method implementation
When an imported Java class has native methods that we expect to implement
from Swift, create a protocol names <Swift Type>NativeMethods that declares
all of the methods that need to be implemented.

From the user perspective, one should make the `@JavaImplements` extension
conform to this protocol. Then, the compiler will ensure that the right
Swift methods exist with the right signatures. Note that the user is
still responsible for ensuring that the appropriate `@JavaImplements` and
`@JavaMethod` annotations are present.