Java2Swift: Drive recursion into nested subclasses from the outside
Rather than having the Java-to-Swift translator walking all nested
classes itself, provide it with the set of nested classes that should
be translated along with the enclosing class. This allows us to be
explicit about the naming of nested classes in the config file (where
we want to), while still getting the convenience of translating the
nested classes automatically for you by default.