Make nested classes explicitly-specifiable and work through some integration issues (#118)
* Java2Swift: Add a test for a generic static method

* Drop the "canonicalClassName" from JavaTypes because it is now incorrect

We're using the "normal" name now that uses $ for nested classes.

* Remove dead code behind a `#if false`

* Teach the Java2Swift plugin to match the nested type naming scheme for Java2Swift

* Regenerate java.util.jar.Attributes with the "isFinal" change

* Java2Swift: Drive recursion into nested subclasses from the outside

Rather than having the Java-to-Swift translator walking all nested
classes itself, provide it with the set of nested classes that should
be translated along with the enclosing class. This allows us to be
explicit about the naming of nested classes in the config file (where
we want to), while still getting the convenience of translating the
nested classes automatically for you by default.

* Update JavaSieve example now that we're handling nested classes

* Java2Swift: Use the parent's Swift type name to form nested type names

* Escape member types named "Type"

* Work around issues with Java classes named "Type" by renaming them "JavaType"

This is deeply unfortunate, but we need compile-side fixes for
nested types named "Type" to work.