Java2Swift: Separate "reference type" from "value type" mappings
While in general we want to map Java classes to Swift value types,
there are places where we need to work with the class type, such as
generic parameters. Identify those places where it is okay to bridge
to a value type (e.g., parameters, results, field types) and do so.

Elsewhere, use the Swift type that directly wraps the Java class. For
example, when dealing with superclasses or with generic arguments to
Java generic classes.

As part of this, provide a class mapping for java.lang.String (as
JavaString) and Java arrays (as the non-generic JavaArray) so we have a
way to refer to such types in positions that cannot use the value type.
Test this out with a few APIs that need it.

Fixes issue #128.