JavaKit: Macro and translator support for mapping Java classes to Swift classes
Java classes are currently translated in Swift structs, and we use is/as
functions to cast up and down the hierarchy. Start down an alternative
path where we translate Java classes into Swift classes, making proper use
of inheritance in Swift to reflect inheritance in Java.

This step updates the @JavaClass macro implementation to check whether
it is being applied to a class (rather than a struct) and adjust its
behavior accordingly:

* The superclass is extracted from the inheritance clause of the class
rather than the "extends" argument to the macro.
* The "full Java class name" static member becomes a "class" member.
For everything other than JavaObject itself, it is an override.
* The `javaHolder` property is only emitted into JavaObject, nowhere else.
* The `init(javaHolder:)` initializer becomes required. For everything
other than JavaObject, it's implemented as a super.init call.
* The `JavaSuperclass` typealias is no longer emitted.

All of this is keyed off the subject of the attribute being a "class"
rather than a "struct", so we keep the existing struct versions working.

Relatedly, extend the translator with an option to map Java classes to
Swift classes. This involves a number of changes:

* Emitting each as an "open class" rather than "public struct"
* Emitting the superclass into the inheritance clause rather than the
"extends" argument
* Emitting methods as "open" rather than "public"
* Only emit methods and fields declared in the class, not inherited ones

This option is only currently only enabled in test cases while we stage
in this functionality.