Java2Swift: Always use the nearest imported superclass for the superclass
When we import a class from Java into Swift, we check whether its superclass
was also imported before generating a reference to that superclass. If it
isn't there, we fell back to JavaObject. That's too conservative, because
there might be a superclass in between that we could use.

Instead, walk up the superclass chain until we find the most-specific
superclass that *is* mapped into Swift, and use that as the generated
superclass. Additionally, use this as the basis for determining when we
need the "override" keyword when in the class-generating mode.