build: adjust the linker flags for JavaKit
`-rpath` is not a universally valid flag for the linker. Adjust the
linker flags to add the library search path on Windows and elide the
`-rpath` flag. It is unclear why there is no `server` subdirectory under
`lib` when using the Oracle JDK 23, but this allows us to find the
necessary library.