Introduce a C type system for lowering purposes
Swift functions are lowered into C-compatible thunks. Those thunks
currently have their types represented in Swift, which is needed for
the `@_cdecl` declaration.

Introduce a representation of the C type system so that we can
describe the specific C types that each parameter has. This
intentionally represents C types in an abstract form that fits will
with the mapping from Swift, for example representing integral types
by the number of bits and their signedness, rather than the actual C
primitive types like `int` or `unsigned long`.

Implement printing of C types and functions, in case we decide that we
want to render C declarations for consumption by humans or tools that
want to bind to them.