[jextract] Prepare all inputs before analyze()
Register all the files to `Swift2JavaTranslator` (and
`NominalTypeResolution`) and `analye()` all the files at once.

`Swift2JavaVisitor.visit(_: ExtentionDeclSyntax)` requires all the
nominal types are prepared in the `nominalResolution`. Otherwise it's
just ignored.

Also previously, writing files happend for each file iteration but
actually, it is not a per-source file operation. Instead, it only writes
the current state of the translator.