[JavaRuntime] Add module.modulemap
Fix a build error in 6.1 toolchain. Apparently some targets have both
`JavaRuntime-tool.build` and `JavaRuntime.build` as the include paths.
The auto generated `module.modulemap` in both directories cause the
error `error: redefinition of module 'JavaRuntime'`. To avoid that,
add `module.modulemap` in the source directory so it's not generated.

Fixes: https://github.com/swiftlang/swift-java/issues/229