jextract: generate one output swift file per input file
Rather than generating an output file per TYPE which we did before.

The per-type mode cannot work correctly because SwiftPM expects to know
all the output Swift files. So we'd have to parse sources and determine
up front what the outputs will be -- doable, but problematic -- instead,
we now generate files based on input files, and map which type goes into
which output file.

This makes it also easier to find where thunks are -- they are in the
same named file as the original type or func was declared in.

This may have some edge case problems still.

Resolves https://github.com/swiftlang/swift-java/issues/365